all modified - ready (v1.0)
