+++ results/trust_scores_over_time.txt
@@ -1,0 +1,100 @@
+0.4289421661348475
+0.4574950578281056
+0.4846203049367004
+0.5103892896898654
+0.5348698252053719
+0.5581263339451039
+0.580220017247848
+0.6012090163854558
+0.6211485655661831
+0.6400911372878745
+0.654578247090147
+0.6645326680689719
+0.6813060346655234
+0.6972407329322482
+0.7123786962856355
+0.7267597614713532
+0.7404217733977854
+0.7466756847278964
+0.752308567158168
+0.7610681388002609
+0.7730147318602474
+0.7843639952672345
+0.7905041288372078
+0.7748955890620136
+0.7742508096089109
+0.7855382691284672
+0.7895613556720438
+0.7964332878884408
+0.7983699568273538
+0.789401458985986
+0.7838813860366864
+0.7946873167348512
+0.7997196175647748
+0.783625303353203
+0.7944440381855438
+0.7929385029429317
+0.8032915777957863
+0.8065186655726639
+0.7938927322940316
+0.785181429012663
+0.7792390242286955
+0.7751187396839283
+0.7688794693663976
+0.7721521625647447
+0.7689862211031744
+0.773011910048014
+0.7628279812122825
+0.7648949154850007
+0.7597085030440827
+0.7487730778918806
+0.7403427573306196
+0.733092286130755
+0.7275710051575511
+0.7191341215663402
+0.708277415488023
+0.695813544713622
+0.6887228674779396
+0.685361724104043
+0.6705103045655081
+0.6650764560038984
+0.6499309665370379
+0.6276177515435202
+0.6051035306330098
+0.5835566874346924
+0.565853853062958
+0.5567194937431428
+0.5403585190559865
+0.5325405931031867
+0.5103385634480279
+0.4945466352756259
+0.4777359701785112
+0.45866583833625235
+0.43886587975277386
+0.42482258576513493
+0.40871478981021103
+0.3921540503197005
+0.37810468113704887
+0.36465778041353
+0.3529415580595198
+0.3423611468232112
+0.33333475614871716
+0.3207680183412816
+0.30780461742421755
+0.29590605321967245
+0.28313575055868884
+0.27222062969742167
+0.26249293154588355
+0.25183495163525627
+0.24165153738682638
+0.23350229385081878
+0.22386051249161107
+0.21579248686703045
+0.2088778625236787
+0.2020506360641616
+0.194423104260954
+0.18701861571457246
+0.1810426849288437
+0.174448884015735
+0.16904310648161505
+0.16381595115753408
