--- telephone_game_app/requirements.txt
@@ -1,23 +1,0 @@
-blinker==1.9.0
-click==8.2.1
-contourpy==1.3.2
-cycler==0.12.1
-Flask==3.1.1
-flask-cors==6.0.0
-Flask-SQLAlchemy==3.1.1
-fonttools==4.59.0
-greenlet==3.2.3
-itsdangerous==2.2.0
-Jinja2==3.1.6
-kiwisolver==1.4.8
-MarkupSafe==3.0.2
-matplotlib==3.10.3
-numpy==2.3.1
-packaging==25.0
-pillow==11.3.0
-pyparsing==3.2.3
-python-dateutil==2.9.0.post0
-six==1.17.0
-SQLAlchemy==2.0.41
-typing_extensions==4.14.0
-Werkzeug==3.1.3
