--- telephone_game_app/src/trust_scores.txt
@@ -1,100 +1,0 @@
-0.52499999999997
-0.5487499999999441
-0.5713125000001235
-0.5927468749999427
-0.6131095312500591
-0.6324540546875138
-0.6508313519532245
-0.6682897843556233
-0.6833652951375975
-0.696060363714112
-0.7088290121951578
-0.7177650615854553
-0.7249659751726562
-0.7348085097474673
-0.7313922509267315
-0.7324968050471036
-0.7269011314614188
-0.7208844082216347
-0.7142293544772396
-0.7014853867533654
-0.7048786174156513
-0.6843580198783397
-0.6600126188843229
-0.6500894879401433
-0.6301891802098138
-0.6054388878660081
-0.5809511101393146
-0.5614535546322806
-0.547341710234081
-0.5249612913890555
-0.5068715601529099
-0.4869029821452519
-0.4714436663713457
-0.4505273163861347
-0.43031178390011665
-0.41361702803846867
-0.39749450996988084
-0.3817697844713594
-0.3653004619144826
-0.3496837721520695
-0.33507958354449935
-0.3223572710339166
-0.308224407482234
-0.2965390204414556
-0.28372123608605343
-0.2718368409484147
-0.26011833223432457
-0.2495140822892745
-0.24006337817481344
-0.23022854259939982
-0.22075794880276908
-0.21223921802929022
-0.20343225712783053
-0.19449731093810493
-0.18733077872453485
-0.18048673978830518
-0.1738815694655581
-0.1668141576589485
-0.15984511644267105
-0.15385452728719892
-0.1479668009228417
-0.1423967942100284
-0.13724195449952925
-0.1323473567745511
-0.12733915560249162
-0.12352469782236651
-0.11902096293124949
-0.11496658145135363
-0.11143991904545214
-0.10761792309317914
-0.10435119360518713
-0.10105613392492697
-0.09751082722868154
-0.0948386192005796
-0.09221002157388504
-0.08976618716185666
-0.08665621113709791
-0.08378673391357602
-0.08139989721789694
-0.079270735690336
-0.07745803223915253
-0.07567513062719487
-0.07404804076250154
-0.07230147205770997
-0.07054306512149107
-0.06910174519874993
-0.06754249127214644
-0.06621953337520495
-0.06504439003977756
-0.064153837204456
-0.06300447867756621
-0.06196842141035445
-0.060650000339836994
-0.059880833656178765
-0.05884095864003661
-0.05793391070803398
-0.05678888183929966
-0.05629693774733449
-0.0554879241933007
-0.05484519465030296
