--- trust_scores.txt
@@ -1,100 +1,0 @@
-0.4803740252900028
-0.5063553240255038
-0.5310375578242286
-0.5544856799330166
-0.5767613959363653
-0.5979233261395472
-0.6180271598325695
-0.6371258018409414
-0.6552695117488947
-0.6725060361614497
-0.6888807343533774
-0.7044366976357076
-0.7192148627539229
-0.7332541196162267
-0.746591413635416
-0.7592618429536434
-0.7710820841392955
-0.7777196465989985
-0.7888336642690477
-0.796566981055597
-0.800646965336149
-0.8106146170693413
-0.8164422195492066
-0.81125344190508
-0.8145657698098261
-0.8086958146526684
-0.8182610239200364
-0.8197646393907004
-0.8249514074211648
-0.8239538370501074
-0.8214728118642693
-0.8281325046043875
-0.8329008793741688
-0.8387641687387942
-0.8407009603018534
-0.8395825789534285
-0.8361451166724234
-0.8443378608388007
-0.8478293011301961
-0.8520295027403534
-0.8519446942700007
-0.8593474595565013
-0.8663800865786764
-0.8625860822497425
-0.8566234448039218
-0.8522172725637258
-0.8458064089355388
-0.8480494218220976
-0.8419052840643236
-0.839226686527776
-0.8398070188680522
-0.84781666792465
-0.8399758345284172
-0.8417020428019955
-0.8306836073285635
-0.8282244269621353
-0.8308132056140295
-0.8334058786666612
-0.8349772513999951
-0.8314950554966608
-0.819186969388495
-0.820635954252403
-0.8214374898731163
-0.810998948712793
-0.8055990012771547
-0.8055773845466299
-0.8008985153192966
-0.7986869228866675
-0.7995442434090003
-0.7918253645718839
-0.7780507630099568
-0.7743565581927905
-0.7745220636164843
-0.771337627102328
-0.7712290790805444
-0.76752595845985
-0.7562913272035252
-0.7488267608433504
-0.7388437561345137
-0.7289265683277889
-0.7269385732447323
-0.7100833112491625
-0.7083624790200381
-0.6913693550690364
-0.6772925539822504
-0.6587195929498053
-0.6391419466356494
-0.6288348493038649
-0.6182264401720055
-0.5935401181634042
-0.5774881122552347
-0.5608970399758068
-0.5454105213103493
-0.5387649952448326
-0.5181184121492575
-0.5090208248751273
-0.48967811696470515
-0.47291921111646923
-0.4598482505606471
-0.44561417136594755
